app/requirements.txt
Flask==0.12.2
Werkzeug==0.12.2
sklearn_deap==0.2.0
seaborn==0.8
xgboost==0.6a2
matplotlib==2.0.2
pandas==0.20.3
scikit_learn==0.19.0
